948 - Fixed bcrhp heritage details print display for sos section
The section was prematurely breaking due to the set display inline-flex, for the browser to break appropriately within the section, its display needs to be set to block.

To prevent the browser clipping text between pages, the clear property needs to be set to none and the display property set to block.